feat(cards): CARD-0039 S2-S4 API contract, ranking consumers, tracker sync
Drop priority from card requests and responses; unknown members on create/edit
are 400. Sort board, orchestrator, Home Next and docs/cards by CardRanking
(ascending). Tracker import maps the 0-5 adapter scale onto importance names;
Linear 1..4 is remapped onto that scale; export labels are priority:critical|high|low.
card.ps1 -Priority is a hard error naming -Importance/-Urgency/-DueAt.

+++ server/Application/Services/TrackerBidirectionalSyncService.cs
@@ -469,7 +469,7 @@ public sealed class TrackerBidirectionalSyncService
         var card = issueRef.Card;
         var desiredStatus = TrackerSyncMarkers.StatusLabel(card.Status);
         var desiredPriority = issueRef.Origin == ExternalIssueOrigin.AntiphonExport
-            ? TrackerSyncMarkers.PriorityLabel((int)card.Importance)
+            ? TrackerSyncMarkers.PriorityLabel(card.Importance)
             : null;
 
         var currentLabels = current.Labels.ToList();
@@ -689,7 +689,7 @@ public sealed class TrackerBidirectionalSyncService
         {
             var labels = BoardService.ParseLabels(card.LabelsJson).ToList();
             labels.Add(TrackerSyncMarkers.StatusLabel(card.Status));
-            if (TrackerSyncMarkers.PriorityLabel((int)card.Importance) is { } p)
+            if (TrackerSyncMarkers.PriorityLabel(card.Importance) is { } p)
                 labels.Add(p);
 
             var body = TrackerSyncMarkers.AppendCardMarkerFooter(
